refactor: Move theming code to separate partial
